Add files via upload
A program designed to help assist trainers in Pokemon GO. I kept forgetting the number and species of pokemon that I had stored in my research stack so rather than simply type it in a notepad, I decided to make a program to keep track of just that. 

Note: keep all background.png, pokestack.ico, and stash.txt in the same folder as the main program.